CDS data request and download

Stage a data request, and optionally download the data to disk. Alternatively you can only stage requests, logging the request URLs to submit download queries later on using wf_transfer.

cds_request(user, request, transfer = TRUE, path = tempdir(),
  time_out = 3600, verbose = TRUE)

Arguments

user
username you used to register the API key when calling cds_set_key. Note: can also be set to NULL to use the .cdsapirc located in your home directory (if existing).

request
nested list with query parameters following the layout as specified on the CDS API page

transfer
logical, download data TRUE or FALSE (default transfer = FALSE)

path
path were to store the downloaded data

time_out
how long to wait on a download to start (default = 3600)

verbose
show feedback on processing

Value

a download query staging url or a netCDF of data on disk

See also

Examples

# Use local keyring:
cds_set_key(user = "1234", key = "abcd1345foo")

#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables

# Specify request, a pressure-level request
# for ERA-5 reanalysis data, temperature 850 hPa.
# One time step, user-defined sub-area, NetCDF.
request_pl <- list(
              "dataset"      = "reanalysis-era5-pressure-levels",
              "product_type" = "reanalysis",
              "format"       = "netcdf",
              "variable"     = "temperature",
              "pressure_level" = "850",
              "year"   = "2000",
              "month"  = "04",
              "day"    = "04",
              "time"   = "00:00",
              "area"   = "70/-20/00/60",
              "format" = "netcdf",
              "target" = "era5-demo.nc")

# Request data/download file using 'user = "1234"'.
# Note: requires that you have stored the login information
# via 'cds_set_key()' (see above).
cds_request(user = "1234",           # user ID (for authentification)
            request = request_pl,    # the request
            transfer = TRUE,         # download the file
            path = ".")              # store data in current working directory

#> Warning: Selecting ‘env’ backend. Secrets are stored in environment variables

#> No encoding supplied: defaulting to UTF-8.

#> Error: <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
#> <html>
#> <head>
#> <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
#> <title>401 Authorization Required</title>
#> </head>
#> <body bgcolor="white">
#> <center><h1>401 Authorization Required</h1></center>
#> <hr>
#> <center>nginx/1.14.0</center>
#> </body>
#> </html>

# Alternatively (as an alternative to use local keyring):
# store user information in .cdsapirc and set user = NULL
# to use login information from .cdsapirc:
cds_request(user = NULL,             # use .cdsapirc
            request = request_pl,    # the request
            transfer = TRUE,         # download the file
            path = ".")              # store data in current working directory

#> - CDS login (from .cdsapirc): user ID: 3258, key = 1fc02******

#> - staging data transfer at url endpoint:

#> - request preparation completed, transfer data

#> - input is a request ID, generate url

#> - Downloading "https://cds.climate.copernicus.eu/api/v2/tasks/542859df-5d0f-4037-9fdf-5ecaadc71da2"

#> - CDS login (from .cdsapirc): user ID: 3258, key = 1fc02******

#> - GET: https://cds.climate.copernicus.eu/api/v2/tasks/542859df-5d0f-4037-9fdf-5ecaadc71da2
#> - GET: https://cds.climate.copernicus.eu/api/v2/tasks/542859df-5d0f-4037-9fdf-5ecaadc71da2
#> Response [http://136.156.132.198/cache-compute-0003/cache/data1/adaptor.mars.internal-1548195695.3489609-25351-11-b7a0aec1-bf50-45be-bcec-fcaab6550f71.nc]
#>   Date: 2019-01-23 00:54
#>   Status: 200
#>   Content-Type: application/x-netcdf
#>   Size: 184 kB
#> <BINARY BODY>

#> - polling server for a data transfer

#> - writing file to disk ("/tmp/RtmpZerMoB/ecmwfr_57e576dd45b2")

#> - move file: /tmp/RtmpZerMoB/ecmwfr_57e576dd45b2 -> ./era5-demo.nc

#> - CDS login (from .cdsapirc): user ID: 3258, key = 1fc02******

#> - request purged from queue!

#> [1] "./era5-demo.nc"

# Second request, surface-level (or single-level) request
# for ERA-5 reanalysis data, total precipitation.
# One time step, user-defined sub-area, NetCDF.
request_sf <- list(
              "dataset"      = "reanalysis-era5-single-levels",
              "product_type" = "reanalysis",
              "format"       = "netcdf",
              "variable"     = "total_precipitation",
              "year"   = "2000",
              "month"  = "04",
              "day"    = "04",
              "time"   = "00:00",
              "area"   = "70/-20/00/60",
              "format" = "netcdf",
              "target" = "era5-demo.nc")

# Surface level request, using '.cdsapirc' login file
cds_request(user = NULL,             # use .cdsapirc
            request = request_sf,    # the request
            transfer = TRUE,         # download the file
            path = ".")              # store data in current working directory

#> - CDS login (from .cdsapirc): user ID: 3258, key = 1fc02******

#> - staging data transfer at url endpoint:

#> - request preparation completed, transfer data

#> - input is a request ID, generate url

#> - Downloading "https://cds.climate.copernicus.eu/api/v2/tasks/9779717d-6b81-4a6e-a91c-fa028894343c"

#> - CDS login (from .cdsapirc): user ID: 3258, key = 1fc02******

#> - GET: https://cds.climate.copernicus.eu/api/v2/tasks/9779717d-6b81-4a6e-a91c-fa028894343c
#> - GET: https://cds.climate.copernicus.eu/api/v2/tasks/9779717d-6b81-4a6e-a91c-fa028894343c
#> Response [http://136.156.132.198/cache-compute-0005/cache/data5/adaptor.mars.internal-1548195731.936494-11076-3-aa5eeacc-0977-465e-9e16-22fec9be896b.nc]
#>   Date: 2019-01-23 00:54
#>   Status: 200
#>   Content-Type: application/x-netcdf
#>   Size: 184 kB
#> <BINARY BODY>

#> - polling server for a data transfer

#> - writing file to disk ("/tmp/RtmpZerMoB/ecmwfr_57e5535fd23d")

#> - move file: /tmp/RtmpZerMoB/ecmwfr_57e5535fd23d -> ./era5-demo.nc

#> - CDS login (from .cdsapirc): user ID: 3258, key = 1fc02******

#> - request purged from queue!

#> [1] "./era5-demo.nc"