List ECMWF Data Store dataset

Returns a list of all ECMWF datasets, covering all Data Store services (i.e. CDS, ADS, CEMS). This function is used to validate the datasets queried by wf_request. For optimization reasons and limit API calls the function is cached and only called once per session (assuming that available products and their information and endpoints aren't updated on a regular sub-daily basis).

wf_datasets(service = c("cds", "ads", "cems"), simplify = TRUE)

Arguments

service

which service to use, one of webapi, cds or ads (default = webapi)

simplify

simplify the output, logical (default = TRUE). When not simplified the raw API return is provided as a nested list, for debugging purposes mostly.

Value

returns a data frame with the ECMWF Data Store datasets

See also

Author

Koen Hufkens

Examples

if (FALSE) { # \dontrun{
# get a list of ECMWF Data Store datasets
wf_datasets()
} # }