ECMWF services list

Returns a list of services

wf_services(email, simplify = TRUE)

Arguments

email
email address used to sign up for the ECMWF data service and used to retrieve the token set by wf_set_key

simplify
simplify the output, logical (default = TRUE)

Value

returns a nested list or data frame with the ECMWF services

See also

Examples

# NOT RUN {
# }
# NOT RUN {
# set key
wf_set_key(email = "test@mail.com", key = "123")

# get a list of services
wf_services("test@mail.com")

# get a list of datasets
wf_datasets("test@mail.com")
# }