ECMWF data transfer function

Returns the contents of the requested url as a (NetCDF) file downloaded to disk or the current status of the requested transfer.

wf_transfer(
  url,
  user = "ecmwfr",
  path = tempdir(),
  filename = tempfile("ecmwfr_", tmpdir = ""),
  verbose = TRUE
)

Arguments

url

R6 wf_request) query output or API endpoint

user

user (email address) used to sign up for the ECMWF data service, used to retrieve the token set by wf_set_key.

path

path were to store the downloaded data

filename

filename to use for the downloaded data

verbose

show feedback on data transfers

Value

a (netCDF) file of data on disk as specified by a wf_request

Details

Normal workflows would use the methods included in returned objects. This is for legacy support and custom scripting only.

See also

Author

Koen Hufkens

Examples

if (FALSE) { # \dontrun{
# request data and grab url and try a transfer
# (request not provided)
r <- wf_request(request, transfer = FALSE)

# check transfer, will download if available
wf_transfer(r$get_url())
} # }