ECMWF data request and download

Stage a data request, and optionally download the data to disk. Alternatively you can only stage requests, logging the request URLs to submit download queries later on using wf_transfer. Note that the function will do some basic checks on the request input to identify possible problems.

wf_request(
  request,
  user,
  transfer = TRUE,
  path = tempdir(),
  time_out = 3600,
  job_name,
  verbose = TRUE
)

Arguments

request
nested list with query parameters following the layout as specified on the ECMWF APIs page

user
user (email address) used to sign up for the ECMWF data service, used to retrieve the token set by wf_set_key

transfer
logical, download data TRUE or FALSE (default = TRUE)

path
path were to store the downloaded data

time_out
how long to wait on a download to start (default = 3*3600 seconds).

job_name
optional name to use as an RStudio job and as output variable name. It has to be a syntactically valid name.

verbose
show feedback on processing

Value

a download query staging url or (invisible) filename of the file on your local disc

See also

Author

Koen Hufkens

Examples

if (FALSE) {
# set key
wf_set_key(user = "test@mail.com", key = "123")

request <- list(stream = "oper",
   levtype = "sfc",
   param = "167.128",
   dataset = "interim",
   step = "0",
   grid = "0.75/0.75",
   time = "00",
   date = "2014-07-01/to/2014-07-02",
   type = "an",
   class = "ei",
   area = "50/10/51/11",
   format = "netcdf",
   target = "tmp.nc")

# demo query
wf_request(request = request, user = "test@mail.com")

# Run as an RStudio Job. When finished, will create a
# variable named "test" in your environment with the path to
# the downloaded file.
wf_request(request = request, user = "test@mail.com", job_name = "test")
}